misc: move `get_all_passes` to xdsl.transforms
