+++ tests/test_cli.py
@@ -1,5 +1,5 @@
 from xdsl.dialects import get_all_dialects
-from xdsl.tools.command_line_tool import get_all_passes
+from xdsl.transforms import get_all_passes
 
 
 def test_get_all_passes_names():
